Extract the four base character types to a new file.
