+++ __init__.py
@@ -1,3 +1,3 @@
-from .bot import Game, Player, Townsfolk, Outsider, Minion, Demon, Traveler, AbilityModifier, NULL_GAME
+from .bot import Game, Player, Traveler, AbilityModifier, NULL_GAME
 
-__all__ = ['Game', 'Player', 'Townsfolk', 'Outsider', 'Minion', 'Demon', 'Traveler', 'AbilityModifier', 'NULL_GAME']
+__all__ = ['Game', 'Player', 'Traveler', 'AbilityModifier', 'NULL_GAME']
