+++ model/__init__.py
@@ -1,4 +1,3 @@
-from .character import Character
 from .script import Script
 
-__all__ = ['Character', "Script"]
+__all__ = ["Script"]
