+++ model/characters/__init__.py
@@ -1,0 +1,4 @@
+from .character import Character
+from .character_types import *
+
+__all__ = ["Character", "Townsfolk", "Outsider", "Minion", "Demon"]
